# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coral Companion API Documentation",
        "description": "Documentation for the API used in the Coral Companion app",
        "contact": {
            "name": "Paula Rumeu",
            "email": "paururo08@hotmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution Share Alike 4.0 International",
            "url": "https://creativecommons.org/licenses/by-sa/4.0/"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://github.com/salem404/coral-companion/raw/master/frontend/src/assets/img/logo-color.svg"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost"
        },
        {
            "url": "https://coral-companion.railway.com/api",
            "description": "Railway"
        }
    ],
    "paths": {
...
        "/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get all items",
                "description": "Get all items from the database",
                "operationId": "64b466e68402b0763a8b7dc3b586a5ad",
                "responses": {
                    "200": {
                        "description": "Success: Return all items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Item"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found: Items don't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No items found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create a new item",
                "description": "Create a new item using the data provided in the request body. (Admin only)",
                "operationId": "38910978a79a2a2fe1af9fb02e683d81",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ItemCreate"
                },
                "responses": {
                    "201": {
                        "description": "Success: Item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Data validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins can create items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to create an item"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get an item",
                "description": "Get an item using the id provided in the request URL",
                "operationId": "3ca89cbe80fb724b4e571eea43f8da7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return an item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found: Item doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item with id 1 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update an item",
                "description": "Update an item by ID in the URL and data provided in the request body. (Admin only)",
                "operationId": "5bee0c1299b966d54c3d44e6fc41b7df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ItemUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Success: Item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Data validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins can update items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to update an item"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found: Item doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item with id 1 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete an item",
                "description": "Delete an item by ID sent in the URL. (Admin only)",
                "operationId": "b2f163bf1735121598620e780afbe870",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Only admins can delete items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized to delete an item"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found: Item doesn't exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item with id 1 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Item": {
                "required": [
                    "name",
                    "id",
                    "icon"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sea Grapes"
                    },
                    "type_id": {
                        "$ref": "#/components/schemas/Type"
                    },
                    "icon": {
                        "type": "string",
                        "example": "sea-grapes.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T17:26:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T17:26:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Type": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "ItemUpdate": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Sea Grapes"
                                },
                                "type_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "icon": {
                                    "type": "string",
                                    "example": "sea-grapes.png"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ItemCreate": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "icon"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Sea Grapes"
                                },
                                "type_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "icon": {
                                    "type": "string",
                                    "example": "sea-grapes.png"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}